structure for fills
