+++ plugin-src/transformers/partials/index.ts
@@ -1,6 +1,7 @@
 export * from './transformBlend';
 export * from './transformChildren';
 export * from './transformDimensionAndPosition';
+export * from './transformFills';
 export * from './transformSceneNode';
 export * from './transformStrokes';
 export * from './transformVectorPaths';
