+++ plugin-src/translators/index.ts
@@ -1,7 +1,5 @@
 export * from './translateBlendMode';
 export * from './translateFills';
-export * from './translateGradientLinearFill';
-export * from './translateSolidFill';
 export * from './translateStrokes';
 export * from './translateTextDecoration';
 export * from './translateTextTransform';
